pyproject.toml
[project]
name = "directors-rota"
version = "1.1.5"
description = "Create director's rota for Phoenix Bridge Club"
authors = [{name = "Jeff", "email" = "<jeffwatkins2000@gmail.com>"}]
readme = "README.md"
requires-python = '>=3.13'

include = [
    "HISTORY.md",
    "src/directors_rota/images/*.png",
    "src/directors_rota/images/*.ico",
    ]

dependencies = [
    "clipboard>=0.0.4",
    "colorama>=0.4.6",
    "dotenv>=0.9.9",
    "icecream>=2.1.4",
    "markdown>=3.8",
    "openpyxl>=3.1.5",
    "psi-toml>=0.0.5",
    "psiconfig>=0.0.13",
    "psiutils>=0.1.63",
    "python-dateutil>=2.9.0.post0",
    "termcolor>=3.1.0",
    "tkinterweb>=4.3.1",
    "workbooky>=0.0.5",
]

[dependency-groups]
dev = ['pytest']

[build-system]
requires = ['uv-build']
build-backend = 'uv_build'
